pyproject.toml
[project]
name = "kosix-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.123.10",
    "google-adk>=1.21.0",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "litellm>=1.80.10",
    "prisma>=0.15.0",
    "psycopg2>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
